xmltypeadapters are now public static inner classes of XmlUtil
