--- src/main/java/address/util/XmlLocalDateAdapter.java
@@ -1,24 +1,0 @@
-package address.util;
-
-import java.time.LocalDate;
-
-import javax.xml.bind.annotation.adapters.XmlAdapter;
-
-/**
- * Adapter (for JAXB) to convert between the LocalDate and the ISO 8601 
- * String representation of the date such as '2012-12-03'.
- * 
- * @author Marco Jakob
- */
-public class XmlLocalDateAdapter extends XmlAdapter<String, LocalDate> {
-
-    @Override
-    public LocalDate unmarshal(String v) {
-        return LocalDate.parse(v);
-    }
-
-    @Override
-    public String marshal(LocalDate v) {
-        return v.toString();
-    }
-}
