--- src/main/java/address/util/XmlLocalDateTimeAdapter.java
@@ -1,20 +1,0 @@
-package address.util;
-
-import javax.xml.bind.annotation.adapters.XmlAdapter;
-import java.time.LocalDateTime;
-
-/**
- * Adapter (for JAXB) to convert between the LocalDateTime and its string representation
- */
-public class XmlLocalDateTimeAdapter extends XmlAdapter<String, LocalDateTime> {
-
-    @Override
-    public LocalDateTime unmarshal(String v) throws Exception {
-        return LocalDateTime.parse(v);
-    }
-
-    @Override
-    public String marshal(LocalDateTime v) throws Exception {
-        return v.toString();
-    }
-}
